AudioSplitter: убрано предупреждение компилятора.
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@4292 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/filters/parser/AudioSplitter/TTAFile.cpp
@@ -114,7 +114,7 @@ HRESULT CTTAFile::Open(CBaseSplitterFile* pFile)
 		return E_FAIL;
 	}
 
-	for (size_t i = 0; i < m_totalframes; i++) {
+	for (int i = 0; i < m_totalframes; i++) {
 		uint32_t size;
 		if (FAILED(m_pFile->ByteRead((BYTE*)&size, 4))) {
 			return E_FAIL;
